Add support for syntax highlighting via Markdown.
Add support for highlighting code blocks, using full
markdown syntax. This makes Blango depend on
python-markdown 1.6b+ and bypass Django's markdown(),
since it currently breaks on markdown 1.7.
mdx_codehilite plugin has been added to the root of
the projects, but Markdown 1.7 and Pygments need
to be provided by the system.